+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is inspired from [Keep a Changelog](http://keepachangelog.com/en/1.0.
 and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- multiple tables per source
+
 ### Fixed
 - temp dir for datapackages
 - query parameters in upload_datapackage